Restore the "natural" order
a) Changed section and variable name percentileValues to postions (as this better reflects the content of the section. Is this ok, or do other modules depend on the exact variable name?
b) Every textbook I know begins with mode median mean in this order. JASP reflects this now also in the output table.
c) Sum is a better fit for section distribution

If this commit is accepted I will reorder further items in this module to make it more beginner friendly.